requirements.txt
biopython
cement
enum34
git+https://github.com/KarrLab/obj_model.git#egg=obj_model-0.0.5
openbabel
setuptools
git+https://github.com/KarrLab/wc_utils.git#egg=wc_utils-0.0.1a5
